requirements.txt
flask
celery
redis
python-dotenv
weaviate-client
langchain
langchain-community
langchain-ollama
llama-index-core
llama-index-embeddings-langchain
llama-index-vector-stores-weaviate
llama-index-readers-file
pypdf
python-docx
python-pptx
openpyxl
pandas
werkzeug
python-magic